+++ Beam/app/Providers/EventServiceProvider.php
@@ -2,8 +2,10 @@
 
 namespace App\Providers;
 
+use App\Observers\GenderBalanceObserver;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+use Remp\BeamModule\Model\Article;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -25,7 +27,6 @@ class EventServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-
-        //
+        Article::observe(GenderBalanceObserver::class);
     }
 }