+++ Beam/config/app.php
@@ -157,6 +157,7 @@ return [
         // App\Providers\BroadcastServiceProvider::class,
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
+        App\Providers\WidgetServiceProvider::class,
     ],
 
     /*