+++ Beam/config/internal.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'gender_balance_enabled' => env('GENDER_BALANCE_ENABLED', false),
+];