+++ Beam/database/migrations/2023_09_07_080805_create_article_meta_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateArticleMetaTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('article_meta', function (Blueprint $table) {
+            $table->id();
+            $table->integer('article_id')->unsigned();
+            $table->string('key');
+            $table->text('value');
+            $table->timestamps();
+
+            $table->foreign('article_id')->references('id')->on('articles');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('article_meta');
+    }
+}