+++ Beam/extensions/beam-module/tests/Feature/API/ConversionsApiTest.php
@@ -18,6 +18,8 @@ class ConversionsApiTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);