+++ Beam/extensions/beam-module/tests/Feature/AggregateArticlesViewsTest.php
@@ -19,6 +19,8 @@ class AggregateArticlesViewsTest extends TestCase
 
     public function testJob()
     {
+        Article::unsetEventDispatcher();
+
         $pageviews1 = <<<JSON
 [
     {