+++ Beam/extensions/beam-module/tests/Feature/DataTables/ArticleConversionsDataTableTest.php
@@ -22,6 +22,8 @@ class ArticleConversionsDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);