+++ Beam/extensions/beam-module/tests/Feature/DataTables/ArticlePageviewsDataTableTest.php
@@ -20,6 +20,8 @@ class ArticlePageviewsDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);