+++ Beam/extensions/beam-module/tests/Feature/DataTables/AuthorsDataTableTest.php
@@ -21,6 +21,8 @@ class AuthorsDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);