+++ Beam/extensions/beam-module/tests/Feature/DataTables/ConversionsDataTableTest.php
@@ -21,6 +21,8 @@ class ConversionsDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);