+++ Beam/extensions/beam-module/tests/Feature/DataTables/TagCategoriesDataTableTest.php
@@ -29,6 +29,8 @@ class TagCategoriesDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);