+++ Beam/extensions/beam-module/tests/Feature/DataTables/TagsDataTableTest.php
@@ -25,6 +25,8 @@ class TagsDataTableTest extends TestCase
     {
         parent::setUp();
 
+        Article::unsetEventDispatcher();
+
         $this->withoutMiddleware([
             VerifyJwtToken::class,
         ]);